task-7/app/requirements.txt
cassandra-driver>=3.25.0
rich>=10.0.0 